organized code with Moralis Params
