+++ frontend/src/Components/HomePage/Home.jsx
@@ -3,52 +3,60 @@ import ImageTest2 from "../../Images/Test02.jpg"
 import ImageTest3 from "../../Images/Test03.png"
 import ImageTest4 from "../../Images/Test04.png"
 import ImageTest5 from "../../Images/Test05.jpg"
+import data from "./HomeData"
+import { useWeb3ExecuteFunction } from "react-moralis";
+
 
 function Home() {
+
+    const contractProcessor = useWeb3ExecuteFunction()
+
+    //function to get the array of objects uploaded in the library from the smart contract
+    const retrieve = async () => {
+        console.log("retireve button clicked")
+        const uploadedCIDS = await _getListOfAllUploadedCIDS()
+        console.log(uploadedCIDS)
+        const metadataArray = []
+        for( let i = 0; i < uploadedCIDS.length; i++){
+            const element = uploadedCIDS[i]
+            const json = await getJSONfiles(element)
+            metadataArray.push(json) 
+        }
+        console.log(metadataArray)
+    }
+
+    const _getListOfAllUploadedCIDS = async () => {
+        const ABI = [{"anonymous":false,"inputs":[{"indexed":false,"internalType":"string","name":"message","type":"string"},{"indexed":false,"internalType":"address","name":"recipient","type":"address"}],"name":"SharewithExisting","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"string","name":"message","type":"string"},{"indexed":false,"internalType":"address","name":"recipient","type":"address"}],"name":"SharewithNew","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"string","name":"message","type":"string"}],"name":"SubsequentUpload","type":"event"},{"anonymous":false,"inputs":[{"indexed":false,"internalType":"string","name":"message","type":"string"}],"name":"Upload","type":"event"},{"inputs":[{"internalType":"address","name":"_address","type":"address"},{"internalType":"string[]","name":"_newCidsToUpload","type":"string[]"}],"name":"_addTwoArrays","outputs":[{"internalType":"string[]","name":"","type":"string[]"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"address","name":"_address","type":"address"},{"internalType":"string[]","name":"_cidsToShare","type":"string[]"}],"name":"_addTwoArraysShared","outputs":[{"internalType":"string[]","name":"","type":"string[]"}],"stateMutability":"nonpayable","type":"function"},{"inputs":[],"name":"_getListOfAllUploadedCIDS","outputs":[{"internalType":"string[]","name":"","type":"string[]"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_address","type":"address"}],"name":"_getListOfUploadedCIDS","outputs":[{"internalType":"string[]","name":"","type":"string[]"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_address","type":"address"}],"name":"_getSharedFiles","outputs":[{"internalType":"string[]","name":"","type":"string[]"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"string[]","name":"_cidsToShare","type":"string[]"},{"internalType":"address","name":"_recipient","type":"address"}],"name":"_shareWithExisting","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"string[]","name":"_cidsToShare","type":"string[]"},{"internalType":"address","name":"_recipient","type":"address"}],"name":"_shareWithNew","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"address","name":"","type":"address"},{"internalType":"uint256","name":"","type":"uint256"}],"name":"_sharedFiles","outputs":[{"internalType":"string","name":"","type":"string"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"string[]","name":"_newCidsToUpload","type":"string[]"}],"name":"_subsequentUpload","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"string[]","name":"_cidsToUpload","type":"string[]"}],"name":"_upload","outputs":[],"stateMutability":"nonpayable","type":"function"},{"inputs":[{"internalType":"address","name":"","type":"address"},{"internalType":"uint256","name":"","type":"uint256"}],"name":"_uploadedCIDS","outputs":[{"internalType":"string","name":"","type":"string"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"allUploadedFiles","outputs":[{"internalType":"string","name":"","type":"string"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"address","name":"_address","type":"address"}],"name":"isAnUploader","outputs":[{"internalType":"bool","name":"","type":"bool"}],"stateMutability":"view","type":"function"},{"inputs":[{"internalType":"uint256","name":"","type":"uint256"}],"name":"upLoaders","outputs":[{"internalType":"address","name":"","type":"address"}],"stateMutability":"view","type":"function"},{"inputs":[],"name":"viewBalance","outputs":[{"internalType":"uint256","name":"","type":"uint256"}],"stateMutability":"view","type":"function"}]
+        const options = {
+            chain: "rinkeby", 
+            contractAddress: "0x653857eCaB380c458206C19142310C1a0cA1683F", 
+            functionName: "_getListOfAllUploadedCIDS", 
+            abi: ABI,
+        }
+        return await contractProcessor.fetch({
+            params: options
+        })
+    }
+
+    const fetchIPFSDoc = async (ipfsHash) => {
+        const url = `https://gateway.moralisipfs.com/ipfs/${ipfsHash}`;
+        const response = await fetch(url);
+        return await response.json();
+      }
+
+    const getJSONfiles = async (CidHashStr) => {
+        return await fetchIPFSDoc(CidHashStr) 
+    }
+
+    // retrieve()
+    
+    //dummy data for the test images used during styling
     const images = [
         ImageTest1, ImageTest2, ImageTest3, ImageTest4, ImageTest5
     ]
-
-    const data = [
-        {
-            image : "",
-            name : "Wizard's first Rule",
-            link : "www.toubvoeejvbfevoofev.com",
-            uploader : "0x72c782vg29r26rv8rv0bvr0r8r2r",
-            status: "private"
-        },
-        {
-            image : "",
-            name : "Revenge of the heckles",
-            link : "www.toubvoeejvbfevoofev.com",
-            uploader : "0x6vew6wv7wve987wevvwvhdvjwdv8wewvS",
-            status: "public"
-        },
-        {
-            image : "",
-            name : "Back to Roots",
-            link : "www.infura.ceewcvev00dv.com",
-            uploader : "0x42nv2442v4v29vrrv8vfw82v2e0v2rvrv2r",
-            status: "private"
-        },
-        {
-            image : "",
-            name : "Heaven can wait",
-            link : "www.ipfs.vewnpweoudd88ddvdudv9ev22e.com",
-            uploader : "0x6vew6wvkjbhci086fgx546435esesszze3dv8wewvS",
-            status:"public"
-        },
-        {
-            image : "",
-            name : "Hoping for the Best",
-            link : "www.toubvoeejvbfevoofev.com",
-            uploader : "0x6vewxyytycxtx6689wevvwvhdvjwdv8wewvS",
-            status: "public"
-        }
-    ]
                                                                                                                                               
     return (
-        <div classname="">
+        <div className="">
             <div className="flex flex-wrap space-x-10 w-[75%]">
                 {data.map((elem, i) => {
                     return (
