+++ frontend/src/Components/HomePage/HomeData.jsx
@@ -1,0 +1,39 @@
+const HomeData = [
+    {
+        image : "",
+        name : "Wizard's first Rule",
+        link : "www.toubvoeejvbfevoofev.com",
+        uploader : "0x72c782vg29r26rv8rv0bvr0r8r2r",
+        status: "private"
+    },
+    {
+        image : "",
+        name : "Revenge of the heckles",
+        link : "www.toubvoeejvbfevoofev.com",
+        uploader : "0x6vew6wv7wve987wevvwvhdvjwdv8wewvS",
+        status: "public"
+    },
+    {
+        image : "",
+        name : "Back to Roots",
+        link : "www.infura.ceewcvev00dv.com",
+        uploader : "0x42nv2442v4v29vrrv8vfw82v2e0v2rvrv2r",
+        status: "private"
+    },
+    {
+        image : "",
+        name : "Heaven can wait",
+        link : "www.ipfs.vewnpweoudd88ddvdudv9ev22e.com",
+        uploader : "0x6vew6wvkjbhci086fgx546435esesszze3dv8wewvS",
+        status:"public"
+    },
+    {
+        image : "",
+        name : "Hoping for the Best",
+        link : "www.toubvoeejvbfevoofev.com",
+        uploader : "0x6vewxyytycxtx6689wevvwvhdvjwdv8wewvS",
+        status: "public"
+    }
+]
+
+export default HomeData
